Separate containers crate
