+++ containers/src/huffman_container.rs
@@ -3,7 +3,7 @@
 use std::collections::BTreeMap;
 use timely::container::PushInto;
 
-use crate::trace::implementations::{BatchContainer, OffsetList};
+use differential_dataflow::trace::implementations::{BatchContainer, OffsetList};
 
 use self::wrapper::Wrapped;
 use self::encoded::Encoded;
@@ -41,8 +41,8 @@ impl<B: Ord + Clone + 'static> PushInto<Vec<B>> for HuffmanContainer<B> {
                 bytes.extend(huffman.encode(item.iter()));
                 self.offsets.push(bytes.len());
             },
-            Err(raw) => { 
-                raw.extend(item); 
+            Err(raw) => {
+                raw.extend(item);
                 self.offsets.push(raw.len());
             }
         }
@@ -150,7 +150,8 @@ impl<B: Ord+Clone> Default for HuffmanContainer<B> {
 
 mod wrapper {
 
-    use crate::IntoOwned;
+    use differential_dataflow::IntoOwned;
+
     use super::Encoded;
 
     pub struct Wrapped<'a, B: Ord> {
@@ -257,7 +258,7 @@ mod huffman {
 
     use std::collections::BTreeMap;
     use std::convert::TryInto;
-    
+
     use self::decoder::Decoder;
     use self::encoder::Encoder;
 
@@ -284,7 +285,7 @@ mod huffman {
         }
 
         /// Decodes the provided bytes as a sequence of symbols.
-        pub fn decode<I>(&self, bytes: I) -> Decoder<'_, T, I::IntoIter> 
+        pub fn decode<I>(&self, bytes: I) -> Decoder<'_, T, I::IntoIter>
         where
             I: IntoIterator<Item=u8>
         {
@@ -320,7 +321,7 @@ mod huffman {
             while let Some((node, level)) = todo.pop() {
                 match node {
                     Node::Leaf(sym) => { levels.push((level, sym)); },
-                    Node::Fork(l,r) => { 
+                    Node::Fork(l,r) => {
                         todo.push((&tree[*l], level + 1));
                         todo.push((&tree[*r], level + 1));
                     },
@@ -348,13 +349,13 @@ mod huffman {
                 }
             }
 
-            Huffman { 
+            Huffman {
                 encode,
                 decode,
             }
         }
 
-        /// Inserts a symbol, and 
+        /// Inserts a symbol, and
         fn insert_decode(map: &mut [Decode<T>; 256], symbol: &T, bits: usize, code: u64) where T: Clone {
             let byte: u8 = (code >> 56).try_into().unwrap();
             if bits <= 8 {
@@ -379,7 +380,7 @@ mod huffman {
         Fork(usize, usize),
     }
 
-    /// Decoder 
+    /// Decoder
     #[derive(Eq, PartialEq, Ord, PartialOrd, Debug, Default)]
     pub enum Decode<T> {
         /// An as-yet unfilled slot.
@@ -395,7 +396,7 @@ mod huffman {
         /// Tests to see if the map contains any invalid values.
         ///
         /// A correctly initialized map will have no invalid values.
-        /// A map with invalid values will be unable to decode some 
+        /// A map with invalid values will be unable to decode some
         /// input byte sequences.
         fn any_void(&self) -> bool {
             match self {
