+++ differential-dataflow/Cargo.toml
@@ -20,10 +20,9 @@ indexmap = "2.1"
 rand="0.4"
 itertools="^0.13"
 graph_map = "0.1"
-bytemuck = "1.18.0"
 
 [dependencies]
-columnar = "0.3"
+columnar = { workspace = true }
 columnation = "0.1.0"
 fnv="1.0.2"
 paste = "1.0"
