+++ differential-dataflow/src/lib.rs
@@ -105,7 +105,6 @@ pub mod collection;
 pub mod logging;
 pub mod consolidation;
 pub mod capture;
-pub mod containers;
 mod into_owned;
 
 /// Configuration options for differential dataflow.
